requirements/base.txt
appdirs==1.4.3
astroid==2.3.1
attrs==19.2.0
backports.csv==1.0.7
black==19.3b0
Click==7.0
colorama==0.4.1
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.6
django-import-export==1.2.0
et-xmlfile==1.0.1
isort==4.3.21
jdcal==1.4.1
lazy-object-proxy==1.4.2
mccabe==0.6.1
odfpy==1.4.0
openpyxl==3.0.0
Pillow==6.2.0
#psycopg2==2.8.3
psycopg2-binary
pylint==2.4.2
pylint-django==2.0.11
pylint-plugin-utils==0.6
pytz==2019.3
PyYAML==5.1.2
six==1.12.0
sqlparse==0.3.0
tablib==0.13.0
toml==0.10.0
typed-ast==1.4.0
wrapt==1.11.2
xlrd==1.2.0
xlwt==1.3.0
celery==4.4.2